requirements.txt
django
gunicorn
google-generativeai
python-dotenv